requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
scikit-learn==1.7.1
pandas==2.0.3
numpy==1.26.4
requests==2.31.0
joblib==1.3.2
scipy==1.11.1
scikit-surprise==1.1.3
mysql-connector-python==8.1.0
python-dotenv==1.0.0
gunicorn==21.2.0